Add casts pointed out by CodeQL analysis.

+++ src/ext/libtess2/Source/bucketalloc.c
@@ -62,7 +62,7 @@ static int CreateBucket( struct BucketAlloc* ba )
 	unsigned char* it;
 
 	// Allocate memory for the bucket
-	size = sizeof(Bucket) + ba->itemSize * ba->bucketSize;
+	size = sizeof(Bucket) + (size_t)(ba->itemSize) * ba->bucketSize;
 	bucket = (Bucket*)ba->alloc->memalloc( ba->alloc->userData, size );
 	if ( !bucket )
 		return 0;
